requirements.txt
Flask
Pillow
torch
torchvision
--extra-index-url https://download.pytorch.org/whl/cu118
